+++ src/azure-cli/HISTORY.rst
@@ -16,6 +16,7 @@ Release History
 **Network**
 
 * network nic ip-config add: Fixes #9861 where --ids was inadvertently exposed but did not work.
+* network application-gateway http-settings create/update: Fixes #9604. Add `--root-certs` to support user associate trusted root certificates with the HTTP settings.*
 
 **RBAC**